feat: Separate parsers into utility script
Because each inference class requires several different command line arguments, and ones like OpticalFlow are used in multiple places, it makes more sense to have a single file with argument parsers in it instead of copy-pasting all the time. Argparse makes this extra easy because the parsers can be stacked using the `parents` keyword argument so they can be mixed and matched as necessary.