refactor: Store optical flow in original size
Originally, the returned optical flow was being resized to the original frame size since that could easily be pulled from the video metadata. However, this meant the optical flow was resized twice which leads to loss of information/resolution and the first resize is unnecessary.

To solve this, I calculate the render size i.e. the size of the returned optical flow to initialize the optical flow array. Ideally, the optical flow inference pipeline wouldn't resize the images but this fixes things for now.