fix: Use RGB frames starting from index 0 instead of 11
The first 10 optical flow frames should be associated with the first RGB frame i.e. the last 10 RGB frames should be unused instead of the first one. A
cyclic numpy array is used to store the frames similar to the way optical flow has been stored in a sliding window array.